vulnerabilities/python/flask/confusion/requirements.txt
flask>=3,<4
pydantic >=2,<3
email-validator>=2,<3
sqlalchemy>=2,<3
psycopg[binary,pool]>=3,<4
bcrypt>=4,<5
pyjwt >=2,<3
